DONE (pretty much)!
All features I planned for the app are implemented and working as
intended.

I’m going to leave in all of my console.log debug statements to help
illustrate the process I used to create the app.

I still have a couple small issues in the CSS that need to be fixed but
I’ll probably save that for later.